Automatically set up remote building on non-Linux hosts
This allows running NixOps on platforms like Mac OS X without
requiring the user to manually set up distributed building.  Instead
we let the machines we provisioned do the builds.

Caveat: there may be a lot of unnecessary I/O.  For instance, it's not
really necessary to copy the closure back to the coordinator.  Also,
since all machines are used for remote building, we may build a
package for machine X on machine Y, copy it from Y to the coordinator,
then from the coordinator to X.