hetzner: Replace hetzner_utils with external lib.
hetzner_util originally was based off that library. As there is a demand for the
library, it doesn't make sense to duplicate efforts so I'm hereby adding it to
the dependencies and use the external lib instead.

Unfortunately this causes a name clash, because Python prefers to use relative
imports over absolute ones, so we need to add something from the __future__.

More details about absolute imports and the mentioned __future__ can be found
here:

http://docs.python.org/2/whatsnew/2.5.html#pep-328-absolute-and-relative-imports

Signed-off-by: aszlig <aszlig@redmoonstudios.org>