Build 64-bit-enabled VirtualBox images.
NixOps, when run on x86_64-linux machines, builds 64-bit guests.
However, on my machine, the resulting kernels hang when booted in the
(32-bit?) VirtualBox environment produced by the "--ostype Linux" call
changed by this patch. Therefore, for the sake of showing what to
change, let's unconditionally build 64-bit images.

(Later on, we can make the architecture of the image dependent on the
architecture of the guest kernel that we hope to run.)