backends: Pass keyword args to get_ssh_flags().
The superclass has a keyword argument (scp=True/False) which is used by
upload_file() and download_file().

So not supporting these arguments is going to fail for these backends
and instead of passing it via a positional argument I'm using **kwargs
so we shouldn't run into the same problem if future kwargs are added to
get_ssh_flags().

This has been reported by @kevincox in an unrelated pull request (#369).

Signed-off-by: aszlig <aszlig@redmoonstudios.org>